Add release-drafter, add matrix in ci, add bean to trie and ahocorasick, add parent to project, remote Pair's interface
